refactor: 1.쿼리문 static으로 변경 2. RowMapper추가 3. ResponseEntity.ok().body리턴 변경
