--- src/main/java/groupware/dao/UserDao.java
@@ -1,34 +1,0 @@
-package groupware.dao;
-
-import groupware.domain.User;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.stereotype.Repository;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.List;
-
-@Repository
-public class UserDao {
-
-    private final JdbcTemplate jdbcTemplate;
-
-    public UserDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public User findById(Long userId){
-        String query = "select * from user where id = ?";
-        List<User> user = jdbcTemplate.query(query, new RowMapper<User>() {
-            @Override
-            public User mapRow(ResultSet rs, int rowNum) throws SQLException {
-                return new User(
-                        rs.getLong("id"),
-                        rs.getString("name")
-                );
-            }
-        }, userId);
-        return user.get(0);
-    }
-}
