--- src/main/java/groupware/mapper/UserRowMapper.java
@@ -1,7 +1,0 @@
-package groupware.mapper;
-
-import groupware.domain.User;
-import org.springframework.jdbc.core.RowMapper;
-
-public interface UserRowMapper extends RowMapper<User> {
-}
