+++ src/main/java/playground/PlaygroundApplication.java
@@ -1,4 +1,4 @@
-package groupware;
+package playground;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -8,5 +8,4 @@ public class PlaygroundApplication {
     public static void main(String[] args) {
         SpringApplication.run(PlaygroundApplication.class, args);
     }
-
 }
