+++ src/main/java/playground/controller/DocumentController.java
@@ -1,13 +1,12 @@
-package groupware.controller;
+package playground.controller;
 
-import groupware.dto.DocumentOutboxResponse;
-import groupware.dto.DocumentRequest;
-import groupware.dto.DocumentResponse;
-import groupware.service.DocumentService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import playground.dto.DocumentOutboxResponse;
+import playground.dto.DocumentRequest;
+import playground.dto.DocumentResponse;
+import playground.service.DocumentService;
 
 import java.util.List;
 
@@ -20,18 +19,18 @@ public class DocumentController {
     @GetMapping("/document/{documentId}")
     public ResponseEntity<DocumentResponse> getDocumentById(@PathVariable Long documentId) {
         DocumentResponse documentDto = documentService.findDocumentBy(documentId);
-        return new ResponseEntity<>(documentDto, HttpStatus.OK);
+        return ResponseEntity.ok().body(documentDto);
     }
 
     @GetMapping("/documents/outbox")
     public ResponseEntity<List<DocumentOutboxResponse>> getDocumentsOutbox(@RequestParam Long userId) {
         List<DocumentOutboxResponse> documentOutboxDtos = documentService.findDocumentsOutbox(userId);
-        return new ResponseEntity<>(documentOutboxDtos, HttpStatus.OK);
+        return ResponseEntity.ok().body(documentOutboxDtos);
     }
 
     @PostMapping("/document")
-    public ResponseEntity<Long> insertDocument(@RequestBody DocumentRequest documentRequest){
+    public ResponseEntity<Long> insertDocument(@RequestBody DocumentRequest documentRequest) {
         Long id = documentService.insertDocument(documentRequest);
-        return new ResponseEntity<>(id, HttpStatus.OK);
+        return ResponseEntity.ok().body(id);
     }
 }
