+++ src/main/java/playground/dto/DocumentOutboxResponse.java
@@ -1,8 +1,8 @@
-package groupware.dto;
+package playground.dto;
 
-import groupware.domain.Document;
 import lombok.Getter;
 import lombok.Setter;
+import playground.domain.Document;
 
 @Getter
 @Setter
