+++ src/main/java/playground/mapper/UserRowMapper.java
@@ -1,0 +1,18 @@
+package playground.mapper;
+
+import org.springframework.jdbc.core.RowMapper;
+import playground.domain.User;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public class UserRowMapper implements RowMapper<User> {
+
+    @Override
+    public User mapRow(ResultSet rs, int rowNum) throws SQLException {
+        return new User(
+                rs.getLong("id"),
+                rs.getString("name")
+        );
+    }
+}
